refactor(material): fix path to Region enum
